==> 04 data_analysis.Rmd <==
---
title: "Independent effects of tree diversity on aboveground woody and soil carbon pools after six years of experimental afforestation"
subtitle: "Appendix S1: Statistical analyses"
author:
  - Reb L. Bryant, Shan Kothari, Jeannine Cavender-Bares, Stephanie J. Curran,
  - Jake J. Grossman, Sarah E. Hobbie, Charlotte Nash, Grace C. Neumiller, Craig R. See
  - "**Journal:** *Ecological Applications*"
bibliography: references.bib
output:
  pdf_document:
    toc: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = "C:/Users/Shan Kothari/Dropbox/FABCarbonProject/ProcessedData")
```

## Introduction

This document is meant to provide a narrative summary to the statistical analyses accompanying the manuscript "Independent effects of tree diversity on aboveground woody and soil carbon pools after six years of experimental afforestation." We show the R output from the models in the manuscript and test the validity of various statistical assumptions underlying our model choices. In cases when these assumptions are violated, we perform robustness checks to show at the very least that the qualitative structure of the models is supported, and to assess the plausible variation in parameter estimates.

Because this document is only meant for the technical purpose of supporting the adequacy of our statistical analyses, we do not in general make much attempt to present output in a polished form. For example, units are generally omitted from axis labels; readers are advised to look to the main text or the metadata accompanying the data repository for them. We note briefly that, as is standard in R Markdown, all figures correspond to the blocks of code that *precede* them, even if those blocks are on the previous page.

The GitHub repository housing this document can be found at [this link](https://github.com/ShanKothari/FAB-carbon-project). A stable archived version can be found at @kothari2024.

## Setting up the workspace

Our first step is to load a number of packages. Although this paper was a collaborative work in which multiple co-authors were working in their own R environments, all results have been checked under the package numbers listed in *R v.4.2.1*.

```{r packages, echo = T, warning=F, message=F}
library(lme4) ## 1.1.35.1
library(lmerTest) ## 3.1.3
library(MuMIn) ## 1.47.5
library(car) ## 3.1.2
library(performance) ## 0.10.9
library(piecewiseSEM) ## 2.3.0
library(MASS) ## 7.3.60.0.1
library(ggplot2) ## 3.4.4
library(emmeans) ## 1.10.0
```

The archived data can be found at the EDI Data Portal (@bryant2023). After setting the working directory to where we saved the data file, we can read it in.

```{r read}
# setwd(".")
carbon_seq<-read.csv("carbon_sequestration.csv")
```

## Relationships among aboveground woody, soil, and fine root carbon

There are three blocks in the FAB1 experiment. Adding a random intercept for block in our models tends to either yield a singular fit or negligible changes in model parameters--the exceptions being in analyses of macroaggregates or soil moisture. (For most other variables, there really does seem to be little evidence that blocks systematically vary; however, three is also a very small number of levels for estimating a random effect variance.) We generally do not show the output of models with the block random effect if it results in a singular fit or in very small changes in parameter values.

One of the first results presented in the paper is the relationship between soil C in 2013 (pre-treatment) and 2019.

```{r soil_C_legacy1, warning = F, message = F}
## here we use initial and final soil C pools (which accounts for bulk density variation)
## but results are similar using the percentages

## one can add a block random intercept here with lmer, adding +(1|block)
## however,  it doesn't change parameter values much
summary(lm(soilC_2019~soilC_2013,data=carbon_seq))

## in general, we plot separate regression lines for each block just to be
## able to check the generality within/among blocks more easily
## naturally, these do not correspond to the 'overall' best-fit line
ggplot(data=carbon_seq,aes(x=soilC_2013,y=soilC_2019,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

```{r soil_C_legacy1_check,fig.height = 8, echo = F}
check_model(lm(soilC_2019~soilC_2013,data=carbon_seq))
```

Despite a couple influential observations that declined in soil C, it's clear that pre-treatment and 2019 soil C are positively correlated. If we want to understand the drivers of carbon *sequestration*, we have good reason to consider the change in soil C pools rather than simply the endpoint in 2019. We might first want to check whether initial soil C determines how much C accumulates.

```{r soil_C_legacy2, warning = F, message = F}
## again these results are robust to alternate ways of analyzing the data
## including considering percentages and adding block random effects
summary(lm(soilC~soilC_2013,data=carbon_seq))
ggplot(data=carbon_seq,aes(x=soilC_2013,y=soilC,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

```{r soil_C_legacy2_check,fig.height = 8, echo = F}
check_model(lm(soilC~soilC_2013,data=carbon_seq))
```

We see the same influential observations as before, but there remains no evidence for a relationship between initial soil C and change in soil C. We can assume that change in soil C is mainly driven by our treatments rather than be legacies of pre-existing soil conditions.

One of the key questions of the manuscript is: does change in soil C correlate with aboveground woody C?

```{r soil_wood, warning = F, message = F}
## here we'll switch to using pools so that aboveground woody and soil
## are in the same units

## note that a block random intercept results in a singular fit
summary(lm(soilC~woodyC,data=carbon_seq))
ggplot(data=carbon_seq,aes(x=woodyC,y=soilC,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

```{r soil_wood_check,fig.height = 8, echo = F}
check_model(lm(soilC~woodyC,data=carbon_seq))
```

It is clear that there is no relationship between the two variables. However, as we report in the paper, if we subset the data to monoculture plots only, we see a significant interaction between species ID and aboveground woody C on soil C.

```{r soil_wood_mono, warning = F, message = F}
summary(aov(soilC~woodyC*monoculture_species,data=carbon_seq))
ggplot(data=carbon_seq[which(!is.na(carbon_seq$monoculture_species)),],
       aes(x=woodyC,y=soilC,color=monoculture_species))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

```{r soil_wood_mono_check,fig.height=8}
check_model(aov(soilC~log(woodyC)*monoculture_species,data=carbon_seq))
## note that although the VIFs here seem astronomical, the
## generalized VIFs implemented with the vif() function in package
## car are much more suitable for multilevel categorical data
vif(aov(soilC~log(woodyC)*monoculture_species,data=carbon_seq))
## moderately high but much less worrisome!
```

Lastly, we look at the third pool we measured: fine root C. First of all, how does fine root C relate to soil C accumulation?

```{r root_soil, warning = F, message = F}
## note that a block random intercept results in a singular fit
summary(lm(soilC~rootC,data=carbon_seq))
ggplot(data=carbon_seq,aes(x=rootC,y=soilC,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

```{r root_soil_check,fig.height = 8, echo = F}
check_model(lm(soilC~rootC,data=carbon_seq))
```

There is fairly weak evidence for a positive relationship between fine root C and soil C accumulation. Next, we check the relationship with aboveground woody C.

```{r root_woody, warning = F, message = F}
## note that a block random intercept results in a singular fit
summary(lm(rootC~woodyC,data=carbon_seq))
ggplot(data=carbon_seq,aes(x=woodyC,y=rootC,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

```{r root_woody_check,fig.height = 8, echo = F}
check_model(lm(rootC~woodyC,data=carbon_seq))
```

Due to the highly skewed distribution of root C, this model evidently violates the normality of residuals assumption of OLS. Although we have good reasons to want to keep root C in its original units, we could try the model with log-transformed root C as a robustness check:

```{r root_soil_log, warning = F, message = F}
## note that a block random intercept results in a singular fit
summary(lm(log(rootC)~woodyC,data=carbon_seq))
ggplot(data=carbon_seq,aes(x=woodyC,y=log(rootC),color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

```{r soil_root_log_check,fig.height = 8, echo = F}
check_model(lm(log(rootC)~woodyC,data=carbon_seq))
```

Although the functional form of the relationship now evidently differs, the coefficient is still negative and the model does somewhat better at meeting the assumptions.

## Changes in soil C in relation to other soil characteristics

We measured three other edaphic properties that we consider here for their relationships with each other and with soil C:

-   macroaggregates (250 $\mu m$ diameter and up)
-   soil moisture
-   pH

First of all, how do these three correlate with each other? Macroaggregates are commonly associated with the ability to retain moisture, so we can test the relationship between macroaggregates and soil moisture.

```{r moisture_macro, warning = F, message = F}
summary(lmer(soil_moisture~macro250+(1|block),data=carbon_seq))
## for F-statistics of mixed-effects models, one can run:
# anova(lmer(soil_moisture~macro250+(1|block),data=carbon_seq))
## and for Nakagawa & Schielzeth Rc2 and Rm2, one can run:
# r.squaredGLMM(lmer(soil_moisture~macro250+(1|block),data=carbon_seq))

## another possibility is to simply treat block
## as a fixed effect to avoid endogeneity problems
## since blocks vary in macroaggregates
## however, this yields very similar results
summary(lm(soil_moisture~macro250+as.factor(block),data=carbon_seq))

ggplot(data=carbon_seq,aes(x=macro250,y=soil_moisture,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

```{r moisture_macro_check,fig.height = 8, echo = F}
check_model(lmer(soil_moisture~macro250+(1|block),data=carbon_seq))
```

There is evidently a close relationship between soil moisture and macroaggregate fraction. How about pH? As it turns out, it isn't correlated with soil moisture but is correlated with macroaggregates.

```{r moisture_pH, warning = F, message = F}
## block random intercept yields singular fit
summary(lm(pH~soil_moisture,data=carbon_seq))

ggplot(data=carbon_seq,aes(x=soil_moisture,y=pH,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

```{r macro_pH, warning = F, message = F}
## block random intercept yields singular fit
summary(lm(pH~macro250,data=carbon_seq))

ggplot(data=carbon_seq,aes(x=macro250,y=pH,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

```{r macro_pH_check,fig.height = 8}
check_model(lm(pH~macro250,data=carbon_seq))
```

These plots make it rather clear that macroaggregates and soil moisture show considerable block-level variation. Indeed, Block 3 has a much lower macroaggregate fraction than the other two, and Blocks 2 and 3 have lower soil moisture than Block 1.

```{r macro_block, echo = F, warning = F, message = F}
ggplot(data=carbon_seq,aes(x=as.factor(block),y=macro250))+
  geom_boxplot()+geom_point()+
  theme_bw()

ggplot(data=carbon_seq,aes(x=as.factor(block),y=soil_moisture))+
  geom_boxplot()+geom_point()+
  theme_bw()
```

Are any of macroaggregates, soil moisture, or pH influenced by tree species richness? Apparently not, as shown below.

```{r soil_richness, warning = F, message = F}
summary(lmer(macro250~species_richness+(1|block),data=carbon_seq))
ggplot(data=carbon_seq,aes(x=species_richness,y=macro250,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
## we note in passing that there appear to be multiple influential observations here
## that may influence these results, but examining the plots it strikes us as unlikely
## that any alternate 'more appropriate' model would return a significant relationship

summary(lmer(soil_moisture~species_richness+(1|block),data=carbon_seq))
ggplot(data=carbon_seq,aes(x=species_richness,y=soil_moisture,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)

summary(lm(pH~species_richness,data=carbon_seq)) ## singular fit with block
ggplot(data=carbon_seq,aes(x=species_richness,y=pH,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

A lot of recent work in ecosystem ecology links the accumulation of stable soil C to the formation of macroaggregates. We could ask: does the percentage of macroaggregates (in 2019) correlate with soil carbon in 2013 (pre-treatment) or 2019? (Here, we put soil C variables on the left side of the formula because we expect the causality to be that macroaggregate formation increases the stability of soil C. This is important because the random intercept terms are quite important. In analyses of soil moisture and pH, we keep soil C variables on the left side even though the causality is less straightforward; however, it is less consequential due to the lack of random intercept terms.)

```{r soil_macro_init, warning = F, message = F}
## note that these relationships would be significant in the absence
## of the block random intercept
summary(lmer(soilC_2013~macro250+(1|block),data=carbon_seq))
# summary(lmer(perC_2013~macro250+(1|block),data=carbon_seq))

## another possibility is to simply treat block
## as a fixed effect to avoid endogeneity problems
## since blocks vary in macroaggregates
## however, this yields the same results
summary(lm(soilC_2013~macro250+as.factor(block),data=carbon_seq))

ggplot(data=carbon_seq,aes(x=macro250,y=soilC_2013,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

```{r soil_macro_final, warning = F, message = F}
## note that these relationships would be significant/marginal in the absence
## of the block random intercept
summary(lmer(soilC_2019~macro250+(1|block),data=carbon_seq))
# summary(lmer(perC_2019~macro250+(1|block),data=carbon_seq))

## another possibility is to simply treat block
## as a fixed effect to avoid endogeneity problems
## since blocks vary in macroaggregates
## however, this yields the same results
summary(lm(soilC_2019~macro250+as.factor(block),data=carbon_seq))

ggplot(data=carbon_seq,aes(x=macro250,y=soilC_2019,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

```{r soil_macro_change, warning = F, message = F}
## when looking at change in soil C, the block effect reaches a singular fit
summary(lm(soilC~macro250,data=carbon_seq))

ggplot(data=carbon_seq,aes(x=macro250,y=soilC,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

These plots offer scant evidence that macroaggregates have much relation to soil C, accounting for the non-independence of plots from the same block. From a more basic soil science perspective, we might wish to see whether there is a correlation between soil percent C and macroaggregates across the full dataset (without block random or fixed effects). There is, although this relationship is perhaps surprisingly weak, and (even more surprisingly) is stronger when examining percent C in 2013 than 2019:

```{r perC_macro}
summary(lm(perC_2013~macro250,data=carbon_seq))
ggplot(data=carbon_seq,aes(x=macro250,y=perC_2013,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)

summary(lm(perC_2019~macro250,data=carbon_seq))
ggplot(data=carbon_seq,aes(x=macro250,y=perC_2019,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

To the extent that there appears to be a relationship across the full dataset, it appears largely driven by the fact that Block 1 has higher macroaggregates than Block 3 in particular, and greater initial and final percent C than Blocks 2 and 3. (These differences in percent C are largely offset by lower bulk density when calculating the top 20 cm C pool, such that there remain no post-treatment block differences in soil C pools or soil C accumulation.)

Although soil C pools and macroaggregates are not strongly related, it turns out that soil C and soil moisture are, based on pools (or percent C) in both 2013 and 2019:

```{r moisture_soilCpool, warning = F, message = F}
## this also yields positive relationships using initial and 2019 percent C
summary(lmer(soilC_2013~soil_moisture+(1|block),data=carbon_seq))
## singular fit
summary(lm(soilC_2019~soil_moisture,data=carbon_seq))

## these findings can easily be reproduced with block as a fixed effect
## with minimal changes to parameter estimates

ggplot(data=carbon_seq,aes(x=soil_moisture,y=soilC_2019,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

```{r moisture_soilCpool_check,fig.height = 8}
check_model(lmer(soil_moisture~soilC_2019+(1|block),data=carbon_seq))
```

However, there is no relationship between soil moisture and soil C accumulation.

```{r moisture_soilC, warning = F, message = F}
## singular fit
summary(lm(soilC~soil_moisture,data=carbon_seq))

ggplot(data=carbon_seq,aes(x=soil_moisture,y=soilC,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

As it turns out, soil pH is positively correlated both with soil C accumulation and (more weakly) with the soil C pool in 2019 (but not in 2013).

```{r soilC_pH, warning = F, message = F}
## block random intercept yields singular fit
summary(lm(soilC~pH,data=carbon_seq))
ggplot(data=carbon_seq,aes(x=pH,y=soilC,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)

## block random intercept yields singular fit
summary(lm(soilC_2013~pH,data=carbon_seq))
ggplot(data=carbon_seq,aes(x=pH,y=soilC_2013,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)

## block random intercept yields singular fit
summary(lm(soilC_2019~pH,data=carbon_seq))
ggplot(data=carbon_seq,aes(x=pH,y=soilC_2019,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
## this model with the 2019 C pool may be
## influenced by a high-leverage point (plot 12)
```

```{r soilC_pH_check,fig.height = 8}
check_model(lm(soilC~pH,data=carbon_seq))
check_model(lm(soilC_2019~pH,data=carbon_seq))
```

We might wonder whether macroaggregates (or soil moisture, or pH) are correlated with fine root C. Once again, the answer is no: we show this for macroaggregates below.

```{r macro_root, warning = F, message = F}
summary(lmer(macro250~rootC+(1|block),data=carbon_seq))
ggplot(data=carbon_seq,aes(x=rootC,y=macro250,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

Lastly, we present a result that we show in the Discussion: plots with higher aboveground woody C have lower soil moisture (and also fewer macroaggregates, although this is wrapped up into the structural equation model below).

```{r woody_soil_moisture}
summary(lmer(soil_moisture~woodyC+(1|block),data=carbon_seq))
ggplot(data=carbon_seq,aes(x=woodyC,y=soil_moisture,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

```{r woody_sm_check,fig.height = 8}
check_model(lmer(soil_moisture~woodyC+(1|block),data=carbon_seq))
```

## Tree biodiversity effects on aboveground woody and belowground C accumulation

One of the core goals of our paper is to show how carbon storage in our experimental plots was influenced by the diversity of planted trees in the first six years of the experiment. We estimated carbon storage across three pools:

-   aboveground wood
-   soil (0-20 cm)--expressed as change from pre-treatment soil carbon
-   fine roots (0-20 cm)

Besides these pools, we also calculated the net biodiversity effect (NBE) for mixture plots as the difference between the observed pools and monoculture-based expectations. (The NBE is synonymous with overyielding, and we use the abbreviation OY in variable names because that was what we chose in early stages of the project.) For aboveground wood, we further partitioned the NBE into complementarity and selection effects following @lh2001.

We considered three facets of diversity:

-   species richness
-   @ll2010's functional dispersion (FDis), as reported by @grossman2017
-   @helmus2007's phylogenetic species variability (PSV), as reported by @grossman2017

The traits that went into the calculation of FDis include: leaf mass per area, leaf nitrogen concentration, leaf habit (categorical), wood density, mycorrhizal type (categorical), leaf calcium concentration, and indices of shade, drought, and waterlogging tolerance from @nv2006. Full details are found in the paper by @grossman2017.

The functional and phylogenetic diversity metrics chosen are designed not to be dependent on species richness. However, FDis and PSV are both 0 for monocultures--by default for FDis and as an analysis choice for PSV. As a result, they do end up somewhat correlated with each other and with species richness. As we will see, multicollinearity does not end up being a severe problem in most cases.

In keeping with the dominant paradigm in biodiversity-ecosystem function (BEF) research, we proposed that planted tree diversity would positively influence all three carbon pools. However, we did not pose hypotheses about which specific facets of tree diversity would matter most. As a result, we used an AICc-based framework to test all models nested within a model that includes species richness, FDis, and PSV. We select the most parsimonious model with a $\Delta AICc<2$ from the model with the lowest AICc. (In effect, this is much like standard null hypothesis significance testing, where $AICc<2$ corresponds to $p>0.05$.) When there are two or more equally parsimonious such models, we selected the one with the lower AICc.

### Aboveground woody carbon

We can start with aboveground woody carbon:

```{r woody_total1, message=F}
mwt<-lm(woodyC~species_richness+FDis+PSV,data=carbon_seq)
mwt_block<-lmer(woodyC~species_richness+FDis+PSV+(1|block),
                data=carbon_seq, REML=F)

options(na.action = "na.fail") # required for dredge to run
mwt_dredge <- dredge(mwt,beta = "none",evaluate = T, rank = AICc)
mwt_block_dredge <- dredge(mwt_block,beta = "none",evaluate = T,rank = AICc)
options(na.action = "na.omit")

## instead of selecting a 'best' model, one could instead perform model averaging
# summary(model.avg(mwt_dredge, subset = delta <= 2))
```

The `dredge` function tests all nested submodels of the specified model. Here we apply it to models both with and without block random intercepts.

We can examine our AIC table, select our 'best' model, and do some basic checks of model assumptions as follows:

```{r woody_total2, fig.height=8}
mwt_dredge ## we'll select the second model
mwt_block_dredge ## parameters are very similar with block
mwt_model<-get.models(mwt_dredge, subset = 2)[[1]]
check_model(mwt_model)
```

A few things are apparent here. First of all, the two retained predictors (FDis and PSV) have opposite-sign coefficients. These predictors have a positive correlation with each other across the full dataset, as shown below:

```{r FDis_PSV, echo=F}
plot(jitter(PSV,factor=50)~jitter(FDis,factor=50),data=carbon_seq)
```

Therefore, their "effects" partly offset each other. However, their variance inflation factor (VIF) is not so high that it would ordinarily raise concerns about multicollinearity.

Perhaps more troubling are the other checks: the posterior predictive check plot, the homoskedasticity plot, and the Q-Q (normality of residuals) plot. It is common for data from BEF experiments not to strictly meet these assumptions of ordinary least-squares regression because the variance is often greatest among monocultures. To assure ourselves that our results are robust, we can use robust regression through iterated re-weighted least squares to reduce the influence of outliers.

```{r woody_total3, message=F}
mwt_robust<-rlm(woodyC~species_richness+FDis+PSV,data=carbon_seq)

options(na.action = "na.fail")
mwt_robust_dredge <- dredge(mwt_robust,beta = "none",evaluate = T,rank = AICc)
options(na.action = "na.omit")

mwt_robust_dredge
```

Here, the 'best' model selected involves the same variables with similar (but not exactly the same) parameter values to when we were using OLS.

### Net biodiversity effect in aboveground woody carbon

Let's repeat the same procedure for the NBE in aboveground woody carbon.

```{r woody_OY, message=F, fig.height=8}
## for the dredge function to work, we need to remove plots with NAs
carbon_seq_woodyOY<-carbon_seq[which(!is.na(carbon_seq$woodyOY)),]

## we'll leave out models with block random intercept
## because they all result in a singular fit
mw<-lm(woodyOY~species_richness+FDis+PSV,
       data=carbon_seq_woodyOY)
## as before, the posterior predictive check and Q-Q plot
## for the OLS model reveal violations of assumptions
mw_robust<-rlm(woodyOY~species_richness+FDis+PSV,data=carbon_seq_woodyOY)

options(na.action = "na.fail")
mw_dredge <- dredge(mw, beta = "none", evaluate = T, rank = AICc)
mw_robust_dredge <- dredge(mw_robust, beta = "none", evaluate = T, rank = AICc)
options(na.action = "na.omit")

mw_dredge ## we select the first model
mw_robust_dredge ## we select the first model

mw_model<-get.models(mw_dredge, subset = 1)[[1]]
check_model(mw_model)
```

Robust regression slightly shrinks parameter estimates here, but yields a qualitatively very similar model.

The NBE is often partitioned via @lh2001's method into complementarity and selection effects. Complementarity effects (CE) are based on the mean of species' *relative* yields and are often taken to indicate synergistic interactions like niche partitioning or facilitation. Selection (SE) effects describe how more productive species tend to show greater (or lesser) relative boosts in productivity than average in mixtures.

Although CE and SE must sum to NBE, it is possible for one to be greater than the NBE (or even total plot biomass) if the other is negative. (It is also possible for total overyielding to be negative, in which case it may be called underyielding.) Although CE and SE are expressed in the same units as NBE (here, kg biomass per ha), it can be hard to make sense of extreme CE or SE values in these terms. For these reasons, we choose to focus on whether they are generally positive or negative than on their actual magnitudes. Starting with the @lh2001 paper itself, it is common for BEF researchers to expect positive CE and non-positive SE on average across plots. This would imply that (1) species tend to overperform in mixture and (2) it is not only the most productive species that overperform in mixture. It may help to look at the data quickly before proceeding.

```{r CESE}
plot(woodyCE~species_richness,data=carbon_seq_woodyOY)
plot(woodySE~species_richness,data=carbon_seq_woodyOY)
```

These plots reveal that there are a couple outliers with highly positive CE and negative SE. The most extreme are plots 14 and 50, both 12-species plots. Because of the odd distribution of CE and SE values, we'll conduct both a parametric t-test and a non-parametric Wilcoxon rank sum test to check whether CE or SE differ from 0. Most research does not concern itself with whether inferences about CE or SE are driven by extreme values, so the t-test is probably more consistent with how results are usually reported. However, the Wilcoxon test may be more appropriate given what we see in the plots above.

```{r CESE_tests}
t.test(carbon_seq_woodyOY$woodyCE)
t.test(carbon_seq_woodyOY$woodySE)
wilcox.test(carbon_seq_woodyOY$woodyCE)
wilcox.test(carbon_seq_woodyOY$woodySE)
```

There is some evidence that CE differs from 0, but not much evidence that SE does.

### Soil C accumulation and net biodiversity effects

We can now do the same for soil C accumulation:

```{r total_soilC, message=F, fig.height=8}
carbon_seq_soilC<-carbon_seq[which(!is.na(carbon_seq$soilC)),]
mst<-lm(soilC~species_richness+FDis+PSV,data=carbon_seq_soilC)

options(na.action = "na.fail")
mst_dredge <- dredge(mst, beta = "none", evaluate = T, rank = AICc)
options(na.action = "na.omit")

mst_dredge ## we pick the first one
mst_model<-get.models(mst_dredge, subset = 1)[[1]]
check_model(mst_model)
```

And for NBE in soil C accumulation:

```{r OY_soilC, message=F}
carbon_seq_soilOY<-carbon_seq[which(!is.na(carbon_seq$soilOY)),]
ms<-lm(soilOY~species_richness+FDis+PSV,data=carbon_seq_soilOY)

options(na.action = "na.fail")
ms_dredge <- dredge(ms, beta = "none", evaluate = T, rank = AICc)
options(na.action = "na.omit")

ms_dredge ## we pick the second one
ms_model<-get.models(ms_dredge, subset = 2)[[1]]
check_model(ms_model)
```

Both of these models seem to meet the assumptions quite well.

### Fine roots and net biodiversity effects

```{r rootC, message=F}
## total root C
carbon_seq_rootC<-carbon_seq[which(!is.na(carbon_seq$rootC)),]
mrt<-lm(rootC~species_richness+FDis+PSV,data=carbon_seq_rootC)

options(na.action = "na.fail")
mrt_dredge <- dredge(mrt, beta = "none", evaluate = T, rank = AICc)
options(na.action = "na.omit")

mrt_dredge ## we select the first model
mrt_model<-get.models(mrt_dredge, subset = 1)[[1]]
check_model(mrt_model)
```

The selected model (with no predictors) does a poor job of meeting assumptions, but even changing out `rootC` for `log(rootC)`, which meets the assumptions better, yields a 'best' model with no predictor variables. The same is true of robust regression. We move on to the NBE.

```{r rootC_OY, message=F, fig.height=8}
## overyielding in root C
carbon_seq_rootOY<-carbon_seq[which(!is.na(carbon_seq$rootOY)),]
mr<-lm(rootOY~species_richness+FDis+PSV,data=carbon_seq_rootOY)
mr_robust<-rlm(rootOY~species_richness+FDis+PSV,data=carbon_seq_rootOY)

options(na.action = "na.fail")
mr_dredge <- dredge(mr, beta = "none", evaluate = T, rank = AICc)
mr_robust_dredge <- dredge(mr_robust, beta = "none", evaluate = T, rank = AICc)
options(na.action = "na.omit")

mr_dredge ## we select the first model
mr_model<-get.models(mr_dredge, subset = 1)[[1]]
check_model(mr_model)
```

Evidently, this model also does a poor job of meeting assumptions. Thus, we show the robust regression results to buttress the claim that the same model is selected even when high-leverage points are not allowed to be as influential.

```{r rootC_OY_robust,message=F}
mr_robust<-rlm(rootOY~species_richness+FDis+PSV,data=carbon_seq_rootOY)

options(na.action = "na.fail")
mr_robust_dredge <- dredge(mr_robust, beta = "none", evaluate = T, rank = AICc)
options(na.action = "na.omit")

mr_robust_dredge
```

PSV is still the only variable selected, but its parameter estimate shrinks quite a bit.

## Tree composition and soil C accumulation

In this section we first test the influence of leaf type and genus on soil C accumulation:

```{r leaf_genus_soilC}
anova(lm(soilC~leaf_type,data=carbon_seq))
## one could run the following to incorporate the block effect
# anova(lmer(soilC~leaf_type+(1|block),data=carbon_seq))
# but it's a singular fit here

## only comparing Quercus and Pinus plots
anova(lm(soilC~major_genus,data=subset(carbon_seq,major_genus!="N")))
```

And the same for root C:

```{r leaf_genus_rootC}
anova(lm(rootC~leaf_type,data=carbon_seq))
anova(lm(rootC~major_genus,data=subset(carbon_seq,major_genus!="N")))
```

These outputs reveal no differences in soil or root C based on leaf type or genus (*Quercus* vs. *Pinus*). However, we do report that both influence macroaggregates:

```{r leaf_genus_macro}
macro_leaf<-lmer(macro250~leaf_type+(1|block),data=carbon_seq)
anova(macro_leaf)
pairs(emmeans(macro_leaf,specs=~leaf_type))
## strong evidence that conifer and deciduous plots differ

macro_genus<-lmer(macro250~major_genus+(1|block),data=subset(carbon_seq,major_genus!="N"))
anova(macro_genus)
```

We can check the diagnostic plots of both models as well:

```{r leaf_genus_macro_check,fig.height=8}
check_model(macro_leaf)
check_model(macro_genus)
```

And the same for soil moisture:

```{r leaf_genus_moisture}
moisture_leaf<-lmer(soil_moisture~leaf_type+(1|block),data=carbon_seq)
anova(moisture_leaf)
pairs(emmeans(moisture_leaf,specs=~leaf_type))
## strong evidence that conifer and deciduous plots differ
## as well as 'both' vs conifer

moisture_genus<-lmer(soil_moisture~major_genus+(1|block),data=subset(carbon_seq,major_genus!="N"))
anova(moisture_genus)
```

```{r leaf_genus_moisture_check,fig.height=8}
check_model(moisture_leaf)
check_model(moisture_genus)
```

Lastly, we explore the effects of mycotype on soil properties. Here, we *do* see an effect on soil C accumulation:

```{r soilC_myco,fig.height=8}
## block effect once again singular
soilC_myco<-lm(soilC~mycotype,data=carbon_seq)
anova(soilC_myco)
pairs(emmeans(soilC_myco,specs=~mycotype))
## no pairwise contrasts are significant

check_model(soilC_myco)
```

But not on macroaggregates or root C:

```{r rootC_myco,fig.height=8}
## block effect once again singular
## log-transforming root C does not change results
rootC_myco<-lm(rootC~mycotype,data=carbon_seq)
anova(rootC_myco)
```

```{r macro_myco,fig.height=8}
macro_myco<-lmer(macro250~mycotype+(1|block),data=carbon_seq)
anova(macro_myco)
```

## Microbial composition and soil C accumulation

We conducted an exploratory analysis of the relationships between various PLFAs (measured in 2016) and either the change in soil C pools or the final pool in 2019. The idea here is that microbial activity could be a driver of how much of the C inputs are actually retained in the soil. However, we don't have the temporal sampling needed to support a clean causal story here, so these correlations are best interpreted as just that: correlations. On a statistical level, it may be worth noting the large deviations from the normality of residuals assumption in some of these models, especially those with a random effect.

Although many PLFAs were measured (and @grossman2019 has further details), we selected certain PLFAs for examination in the analyses below.

```{r 15:0_iso}
summary(lmer(soilC~X15.0_iso.pp+(1|block),data=carbon_seq))
summary(lmer(soilC_2019~X15.0_iso.pm+(1|block),data=carbon_seq))

ggplot(data=carbon_seq,aes(x=X15.0_iso.pm,y=soilC_2019,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

```{r 15:0_iso_check, fig.height=8}
check_model(lmer(soilC_2019~X15.0_iso.pm+(1|block),data=carbon_seq))
```

```{r 17:0_anteiso}
summary(lmer(soilC~X17.0_anteiso.pp+(1|block),data=carbon_seq))
summary(lmer(soilC_2019~X17.0_anteiso.pm+(1|block),data=carbon_seq))

ggplot(data=carbon_seq,aes(x=X17.0_anteiso.pm,y=soilC_2019,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

```{r 17:0_anteiso_check, fig.height=8}
check_model(lmer(soilC_2019~X17.0_anteiso.pm+(1|block),data=carbon_seq))
```

```{r 18.0_10me.pm}
## block yields singular fit here
summary(lm(soilC~X18.0_10me.pp,data=carbon_seq))
summary(lm(soilC_2019~X18.0_10me.pm,data=carbon_seq))

ggplot(data=carbon_seq,aes(x=X18.0_10me.pm,y=soilC_2019,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

```{r X18.0_10me.pm_check, fig.height=8}
check_model(lm(soilC_2019~X18.0_10me.pm,data=carbon_seq))
```

```{r 18.2_w6.9c.pm}
## block yields singular fit for the first of these
summary(lm(soilC~X18.2_w6.9c.pp,data=carbon_seq))
summary(lmer(soilC_2019~X18.2_w6.9c.pm+(1|block),data=carbon_seq))

ggplot(data=carbon_seq,aes(x=X18.2_w6.9c.pp,y=soilC,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

```{r X18.2_w6.9c.pp_check, fig.height=8}
check_model(lm(soilC~X18.2_w6.9c.pp,data=carbon_seq))
```

Besides these taxon-specific PLFAs, we were interested in showing whether total PLFA concentration correlated with soil C or macroaggregates. It does not:

```{r Biomass.pm}
## singular fit for the first of these
summary(lmer(soilC~Biomass.pm+(1|block),data=carbon_seq))
summary(lmer(soilC_2019~Biomass.pm+(1|block),data=carbon_seq))
summary(lmer(macro250~Biomass.pm+(1|block),data=carbon_seq))
```

However, the percent of fungal PLFAs has a marginally significant correlation with soil C accumulation.

```{r fungal_pct}
## singular fit for the first of these
summary(lmer(soilC~gen_fungi_pct+(1|block),data=carbon_seq))

ggplot(data=carbon_seq,aes(x=gen_fungi_pct,y=soilC,color=as.factor(block)))+
  geom_point(size=2)+geom_smooth(method="lm",se=F,linewidth=2)
```

```{r fungal_pct_check, fig.height=8}
check_model(lmer(soilC~gen_fungi_pct+(1|block),data=carbon_seq))
```

## Structural equation modeling of aboveground woody and soil C

We use structural equation modeling (SEM) to tie together our hypotheses in a more integrative manner. Although we found no evidence elsewhere for a correlation between aboveground woody and soil C accumulation, SEMs allow us to directly test whether the influence of species richness on soil C accumulation is mediated by aboveground woody C.

We select six variables:

-   species richness (exogenous)
-   percent of planted trees associated with AM fungi (exogenous)
-   percent of planted trees that are conifers (exogenous)
-   aboveground woody C (endogenous)
-   macroaggregates (endogenous)
-   soil C accumulation (endogenous)

The first step is to drop rows with NAs in any of these variables, then z-standardize all of them so that regression coefficients in the SEM are standardized:

```{r SEM_prep}
carbon_seq_sub<-carbon_seq[-which(is.na(carbon_seq$soilC) | is.na(carbon_seq$macro250)),]

carbon_seq_standard<-carbon_seq_sub
standard_cols<-c("species_richness","woodyC","soilC",
                 "macro250","percentAM","percentCon")
carbon_seq_standard[,standard_cols]<-scale(carbon_seq_standard[,standard_cols])
```

As explained in the main text and Appendix S2, we proposed an initial model based on *a priori* hypotheses and earlier data analyses. The structure of this model is as follows:

```{r SEM_orig, warning=F, message=F}
localfit_orig_model<-psem(
  lmer(woodyC~species_richness+percentCon+(1|block),data=carbon_seq_standard),
  lmer(macro250~percentAM+percentCon+woodyC+(1|block),data=carbon_seq_standard),
  lmer(soilC~species_richness+percentAM+percentCon+woodyC+macro250+(1|block),
     data=carbon_seq_standard)
)

summary(localfit_orig_model)
```

By running `check_models()` on each subcomponent of the model, it becomes clear that percent conifer and aboveground woody C have moderately high VIF due to their high correlation with each other. We removed aboveground woody C from the model for macroaggregates and percent conifer from the model for soil C accumulation, although we caution that in any case it is a challenge to attribute anything definitely to one or the other variable.

Two other things are apparent. One is that the block random effect for soil C reaches a singular fit, so we remove it. The other is that the tests of directed separation show conditional dependence between percent AM and woody C. Since percent AM is exogenous, we add it as a predictor of woody C. We arrive at the following model:

```{r SEM_full, warning=F}
localfit_full_model<-psem(
  lmer(woodyC~species_richness+percentCon+percentAM+(1|block),data=carbon_seq_standard),
  lmer(macro250~percentAM+percentCon+(1|block),data=carbon_seq_standard),
  lm(soilC~species_richness+percentAM+woodyC+macro250,
     data=carbon_seq_standard)
)

summary(localfit_full_model)
```

As the Fisher's C *p*-value \> 0.05 shows, this model fits better than the one above. The in-depth interpretation of this model is in the main text. We can check each of the submodels of the full SEM to make sure that it means standard assumptions.

```{r SEM_checking,fig.height=8}
check_model(lmer(woodyC~species_richness+percentCon+percentAM+(1|block),data=carbon_seq_standard))
check_model(lmer(macro250~percentAM+percentCon+(1|block),data=carbon_seq_standard))
check_model(lm(soilC~species_richness+percentAM+woodyC+macro250,data=carbon_seq_standard))
```

Each of these plots looks reasonably good, barring violations of the normality of residuals assumption, which shouldn't be too concerning. There is no particular need to fit a more minimal model that only retains significant terms, but if we did, it would look like this:

```{r SEM_min, warning=F}
localfit_min_model<-psem(
  lmer(woodyC~species_richness+percentCon+percentAM+(1|block),data=carbon_seq_standard),
  lmer(macro250~percentCon+(1|block),data=carbon_seq_standard),
  lm(soilC~species_richness+percentAM,
     data=carbon_seq_standard)
)
```

## References
